Associations: Added basic acceptance spec
I had to add some weired autoloading magic for the required fixtures of
the spec. Without that we would end up with some nice circular
dependencies. As far as I can tell it at the moment this is the reason
it will work in a Rails project. But this **not** yet confirmed.